# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "iOS Chat Agent Template Backend",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Conversation Status",
        "operationId": "get_conversation_status_api_conversations__conversation_id__get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Chat Response",
        "operationId": "create_chat_response_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Chat Stream",
        "operationId": "create_chat_stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "example": "event: conversation.started\ndata: {\"conversation_id\": \"demo-conversation\", \"agent_id\": \"default\", \"session_history_limit\": 40}\n\nevent: message.delta\ndata: {\"delta\": \"Hello\"}\n\nevent: agent.updated\ndata: {\"agent_name\": \"Planning Specialist\"}\n\nevent: run.item\ndata: {\"name\": \"tool_called\", \"item_type\": \"tool_call_item\"}\n\nevent: message.completed\ndata: {\"conversation_id\": \"demo-conversation\", \"agent_id\": \"default\", \"final_agent_name\": \"Planning Specialist\", \"response_id\": \"resp_demo_123\", \"content\": \"Hello streamed\", \"session_history_limit\": 40}\n\nevent: done\ndata: {}\n\n"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatCompletionRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Message"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatCompletionRequest"
      },
      "ChatCompletionResponse": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "final_agent_name": {
            "type": "string",
            "title": "Final Agent Name"
          },
          "response_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Id"
          },
          "message": {
            "$ref": "#/components/schemas/ChatMessagePayload"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "agent_id",
          "final_agent_name",
          "message"
        ],
        "title": "ChatCompletionResponse"
      },
      "ChatMessagePayload": {
        "properties": {
          "role": {
            "type": "string",
            "const": "assistant",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessagePayload"
      },
      "ConversationStatusResponse": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "exists": {
            "type": "boolean",
            "title": "Exists"
          },
          "session_history_limit": {
            "type": "integer",
            "title": "Session History Limit"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "exists",
          "session_history_limit"
        ],
        "title": "ConversationStatusResponse"
      },
      "ErrorPayload": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorPayload"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status"
          },
          "default_agent_id": {
            "type": "string",
            "title": "Default Agent Id"
          },
          "available_agent_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Agent Ids"
          },
          "session_history_limit": {
            "type": "integer",
            "title": "Session History Limit"
          },
          "agent_config_version": {
            "type": "string",
            "title": "Agent Config Version"
          }
        },
        "type": "object",
        "required": [
          "status",
          "default_agent_id",
          "available_agent_ids",
          "session_history_limit",
          "agent_config_version"
        ],
        "title": "HealthResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}